Create Sass-Store Data with Sass Variables.md
I created a variable in sass called `text-color` and I set it to red. Using this variable I changed the colour of the blog post classes and h2 header to red. *Note: The syntax is vary important the variable starts with a dollar sign, after the name is finished there is a colon at the end and lastly there needs to be a semi colon at the end after css is stated for example. To create a variable that holds the colour red the syntax is: `$variable-name: red;`.