Create Sass-Use @for to Create a Sass Loop.md
I created a for loop using SASS. Iterates though the p elements with the class text-1 to text-5. The for loop is using #j as the parameter. First the class is called with a period( . ) next the class name in this case "text-" lastly the number going form 1 to 5. The default font size is set to 15px for 1st p element. depending on the number the loop multiplies the font size by the number represented by #J. So it go's 15, 30, etc until it reach's 5.