Create Sass-Apply a Style Until a Condition is Met with @while.md
This was very similar to the for directive/loop. First I created a variable called x and gave it the value of 1. Next I used the while directive and give it a condition to keep running while x is less than or equal to 5. Next I used the period followed by the class name with $x and lastly I increment x by 1 each loop.