Create React-Pass Props to a Stateless Functional Component.md
I created a property for my CurrentDate component inside of my calendar component and gave it the value of the JavaScript object`Date()`. This allows me to use this prop in the CurrentDate component. I put the date in a P tag and I accessed it by calling the prop parameter followed by period and name of property.