Create React-Pass an Array as Props.md
This was similar to the previous challenge but instead of passing a object into a prop I put an array there. I put 2 elements in an array inside the first list under the h2 tag Today. Then I put 3 elements into an array and passed it to the list under the Tomorrow h2 tag. Lastly, I return the tasks in the list component. I place them in a p tag and use JavaScript method `.join()` to take them out the array and split them by a space and comma.