Create React-Create a Controlled Input.md
First I created a `handlechange()` function inside of the component `ControlledInput`. The function changes the state property input, by using `setState()` and passing it a function that changes the input to the `event.target.input`.

The `event.target.input` is inside of the input tag at the bottom above the h4 tag. When a user inputs data the `onChange` method changes the state input value by calling `this.handleChange`. The value method is what shows in real time what text is typed in the input.

Lastly, I binded `.this.` to `handleChange` by adding this in the constructor. `  this.handleChange = this.handleChange.bind(this)`