Create React-Pass State as Props to Child Components.md
I gave my Navbar component the name from the state in my app by adding it as a child component of MyApp. I done this by adding `<Navbar name={this.state.name} />` in MyApp component and then I rendered it by placing it in my Navbar component by adding `{this.props.name}` in a JSX tag.