Create React-Add Event Listeners.md
I attached an event listener function inside of my `componentDidMount()` method. I have it 2 parameters the first one is the event in this case `"keydown"` the event needs to be inside quotes. Then I added the callback as the 2nd argument which was `handleKeyPress` this was written as `document.addEventListener("keydown", this.handleKeyPress)`. The 2nd argument needed the `.this` keyword as it is looking for that method inside of the component without `.this` its restricted to only the information inside the `componentDidMount()` method.

Next, I added the `document.removeEventListener()`inside of the `componentWillUnmount()` to remove the event listener it took the same parameters as the `componentDidMount()` method.  The events need to removed after as they may create bugs in other parts of the react application.