Create React-Use Advanced JavaScript in React Render Method.md
There was 2 things I done. First I changed the 52nd line the one with` const answer` I made it equal to`randomIndex`. The `randomIndex` was stored in the state of the component so to access it have had to write `this.state.randomIndex`.

Lastly I placed the answer const inside at the bottom of the code under the h3 and P elements to render out a random string from an array inside the component.