Create React-Render with an If-Else Condition.md
I created an if statement under the render method in MyComponent. I'm testing if the display value is true or false. The value is held in the components state so to access it I need to use `this.state.display`. If the value is true a h1 is shown that has the text  "Displayed!", when it's false this does not show.