Create React-Use && for a More Concise Conditional.md
I used the `&&` logical operator to check if `this.state.display` is true. If it's true `<h1>Displayed!</h1>` is shown if its false then it is not shown.  This is a very useful way to test conditions and render JSX, the same can be done with if/else statements, however, it would be more verbose.