Create React-Use a Ternary Expression for Conditional Rendering.md
First, I initialize the state for the CheckUserAge component and gave it a input and userAge props set to empty stings.

Next, I used the ternary expression to display different buttons depending on the input and when they click submit. The first button was the submit button, this should be shown when `this.state.userAge` is still an empty string. Next was buttonTwo this is shown when the user enters an age over 18. Lastly buttonThree shows when the user enters a age under 18.

<u>example of ternary expression under solution</u>