Create Redux-send Action Data to the Store.md
First I finished off writing the `addNoteText()`function, so that it returns an action object with a property of `type` and a value of `ADD_NOTE` and another prop of `text`and a value of `note`.

Next I finished the `notesReducer` switch statement by a case where `ADD_NOTE`returns `action.text`.