Create Redux-Write a Counter with Redux.md
I created a simple counter with redux. First I set up the action types`INCREMENT`and`DECREMENT` setting them to strings with the same name as constants.

Next I created the reducer for the counter, `counterReducer` has a state set to 0 and an action as parameters. Next I gave it a switch statement that has 2 cases one for increment and the other for decrement, when they are called they return the state incremented or decremented else the state is returned unchanged.

Then I set up the action creators for the increment and decrement, `incAction` returns a object with a type and value of "INCREMENT", `decAction`is the same thing but for the type value is "DECREMENT".

Lastly, I set up the redux store by setting the const store to `Redux.createStore(counterReducer)`.