Create Redux-Remove an Item from an Array.md
I finished off the switch statement in the`immutableReducer` for case "REMOVE_ITEM" I return an array with 1 number deleted. I do this with the stead operator and `.slice()`.

This was bit confusing. The goal is to get rid of the number at the index and return the rest of the array.  For example if the state was `[0,1,2,3,4,5]`and the index is 4 then 4 should be removed at the returned array should be `[ 0, 1, 2, 3, 5 ]` This is shown below in `consol.log `'s.