Create React and Redux Extract State Logic to Redux.md
First I created a action type called "ADD". Next, I created an action creator called "addMessage" that returns an object with a type set to ADD and message set to the message passed in. Then, I created a reducer called "MessageReducer" that handles different action.type 's. Lastly, I created the a store and set it to the redux store with the message reducer passed in.